Replace build system with Maven, remove Gradle - ~30% speed up
For some reason, Gradle is noticeably slower on my system as compared to
Maven, even with a nearly identical project definition. Maven is about
30% faster or so, 2.4 seconds or less compared to Gradle at 3.5 seconds
or there abouts. Only an approximately one second difference but it
feels a lot slower when building with Gradle. Used to be 2x slower but
sped it up by removing one of the unnecessary default tasks, still
uncomfortably slow for my tastes. See console logs for comparison at
https://gist.github.com/deathcap/b6d787435b560b281b03

Also, the Bukkit dependency builds with Maven too, so it is more
consistent if we also use it, instead of Gradle. Oh well.